Change type of idUser from long to int.

+++ CoreService-backend/Controllers/Users.cs
@@ -43,7 +43,7 @@ namespace CoreService_backend.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public ActionResult<UserDto> GetUser(long userId)
+        public ActionResult<UserDto> GetUser(int userId)
         {
             var userDto = _usersService.GetUserById(userId);
